backend+frontend transactions, holdings & currencies

+++ backend/src/main/java/com/example/trading_sim/controllers/WalletController.java
@@ -1,7 +1,9 @@
 package com.example.trading_sim.controllers;
 
+import com.example.trading_sim.models.Holding;
 import com.example.trading_sim.models.User;
 import com.example.trading_sim.models.Wallet;
+import com.example.trading_sim.services.HoldingService;
 import com.example.trading_sim.services.UserService;
 import com.example.trading_sim.services.WalletService;
 import org.springframework.http.HttpStatus;
@@ -12,18 +14,23 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 
 @RestController
 @RequestMapping("/api/wallets")
 public class WalletController {
     private final WalletService service;
+    private final HoldingService holdingService;
     private final UserService userService;
 
-    WalletController(WalletService service, UserService userService)
+    WalletController(WalletService service, HoldingService holdingService, UserService userService)
     {
         this.service = service;
+        this.holdingService = holdingService;
         this.userService = userService;
     }
 
@@ -66,6 +73,26 @@ public class WalletController {
 
         wallet.setId(null);
         wallet.setUserId(user.get().getId());
+        Wallet savedWallet = service.saveByUserId(wallet, user.get().getId());
+
+        return new ResponseEntity<>(savedWallet, HttpStatus.CREATED);
+    }
+
+    @PostMapping("/{id}/reset")
+    @ResponseBody
+    public ResponseEntity<Wallet> reset(@PathVariable Integer id, @AuthenticationPrincipal UserDetails details)
+    {
+        Optional<User> user = userService.findByUsername(details.getUsername());
+        if (user.isEmpty())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        Optional<Wallet> walletOpt = service.findByIdAndUserId(id, user.get().getId());
+        if (walletOpt.isEmpty())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        Wallet wallet = walletOpt.get();
+        wallet.setBalance(10000f);
+        holdingService.deleteAllByWalletId(wallet.getId());
         Wallet savedWallet = service.save(wallet);
 
         return new ResponseEntity<>(savedWallet, HttpStatus.CREATED);
@@ -79,13 +106,12 @@ public class WalletController {
         if (user.isEmpty())
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
-        boolean exists = service.existsByIdAndUserId(id, user.get().getId());
-        if (!exists)
+        if (!service.existsByIdAndUserId(id, user.get().getId()))
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
         wallet.setId(id);
         wallet.setUserId(user.get().getId());
-        Wallet savedWallet = service.save(wallet);
+        Wallet savedWallet = service.saveByUserId(wallet, user.get().getId());
 
         return new ResponseEntity<>(savedWallet, HttpStatus.NO_CONTENT);
     }
@@ -103,6 +129,7 @@ public class WalletController {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
         Wallet wallet = walletOpt.get();
+        String oldName = wallet.getName();
 
         updates.forEach((key, value) -> {
             switch (key) {
@@ -113,14 +140,18 @@ public class WalletController {
             }
         });
 
-        Wallet saved = service.save(wallet);
+        Wallet saved;
+        if (!Objects.equals(oldName, wallet.getName()))
+            saved = service.saveByUserId(wallet, user.get().getId());
+        else
+            saved = service.save(wallet);
+
         return new ResponseEntity<>(saved, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     @ResponseBody
-    public ResponseEntity<Wallet> deleteById(@PathVariable Integer id, @AuthenticationPrincipal UserDetails details)
-    {
+    public ResponseEntity<Wallet> deleteById(@PathVariable Integer id, @AuthenticationPrincipal UserDetails details) {
         Optional<User> user = userService.findByUsername(details.getUsername());
         if (user.isEmpty())
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -128,8 +159,11 @@ public class WalletController {
         if (!service.existsByIdAndUserId(id, user.get().getId()))
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
-        service.deleteById(id);
+        // cant delelte last wallet of user
+        if (service.countByUserId(user.get().getId()) <= 1)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 
+        service.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
